requirements.txt
# WhatsApp SLA Bot Dependencies

# Web Framework
Flask>=3.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests (untuk WhatsApp API calls)
requests>=2.31.0

# Database
PyMySQL>=1.1.0
SQLAlchemy>=2.0.0

# Redis (untuk caching dan queue)
redis>=5.0.0

# Production Server
gunicorn>=21.0.0

# Utilities
python-dateutil>=2.8.0
